insurance/smart_home/requirements.txt
fastapi
uvicorn
pandas
pydantic
requests
streamlit
matplotlib
scikit-learn
paho-mqtt
flask
streamlit-authenticator
python-dotenv
pyjwt